@@ -6,13 +6,13 @@ version: 0.1.0
 homepage: https://github.com/james-alex/gesture_callbacks
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_object_map: ^0.1.1
+  flutter_object_map: ^1.0.0
 
 dev_dependencies:
   pedantic: ^1.9.2
