@@ -21,13 +21,12 @@ class GestureCallbacks extends JoinableObject<GestureCallbacks> {
     this.onSecondaryLongPress,
     this.onHover,
     bool inherit = true,
-  })  : assert(inherit != null),
-        super(inherit: inherit);
+  }) : super(inherit: inherit);
 
   /// A unique identifier used to define the retrieve global [GestureCallbacks]
   /// with the [GestureCallbacks.global] constructor/provider and the related
   /// static methods.
-  final Key key;
+  final Key? key;
 
   /// Called when a tap with a primary button has occurred.
   final GestureTapCallback onTap;
@@ -48,12 +47,12 @@ class GestureCallbacks extends JoinableObject<GestureCallbacks> {
   final GestureTapCallback onSecondaryLongPress;
 
   /// Called when a pointer enters or exits a widget's area.
-  final ValueSetter<bool> onHover;
+  final ValueSetter<bool>? onHover;
 
   /// Returns a new [GestureCallbacks] by combining `this` and [callbacks]
   /// respective values into new functions that call the values of both.
   @override
-  GestureCallbacks combine(GestureCallbacks callbacks) {
+  GestureCallbacks combine(GestureCallbacks? callbacks) {
     if (!inherit || callbacks == null) {
       return this;
     }
@@ -74,7 +73,7 @@ class GestureCallbacks extends JoinableObject<GestureCallbacks> {
   /// Returns a new [GestureCallbacks] containing `this` object's values,
   /// where any `null` values fallback to [callbacks]` values.
   @override
-  GestureCallbacks merge(GestureCallbacks callbacks) {
+  GestureCallbacks merge(GestureCallbacks? callbacks) {
     if (!inherit || callbacks == null) {
       return this;
     }
@@ -99,8 +98,6 @@ class GestureCallbacks extends JoinableObject<GestureCallbacks> {
   /// accessed by the [global] callbacks constructor/provider and
   /// related methods.
   void registerAsGlobal<T>([JoinMethod join = JoinMethod.combine]) {
-    assert(join != null);
-
     _global.add<T>(this, key: key, join: join);
   }
 
@@ -120,20 +117,20 @@ class GestureCallbacks extends JoinableObject<GestureCallbacks> {
   /// If [joinDynamic] isn't `null` and global callbacks associated with the
   /// given [key] and a [dynamic] sub-type exist, they will be [merge]d or
   /// [combine]d with the returned callbacks.
-  static GestureCallbacks global<T>({
-    Key key,
-    GestureTapCallback onTap,
-    GestureTapCallback onDoubleTap,
-    GestureTapCallback onLongPress,
-    GestureTapCallback onSecondaryTap,
-    GestureTapCallback onSecondaryLongPress,
-    ValueSetter<bool> onHover,
+  static GestureCallbacks? global<T>({
+    Key? key,
+    GestureTapCallback? onTap,
+    GestureTapCallback? onSecondaryTap,
+    GestureTapCallback? onDoubleTap,
+    GestureLongPressCallback? onLongPress,
+    GestureLongPressCallback? onSecondaryLongPress,
+    ValueSetter<bool>? onHover,
     JoinMethod join = JoinMethod.combine,
-    JoinMethod joinDynamic = JoinMethod.combine,
+    JoinMethod? joinDynamic = JoinMethod.combine,
   }) {
-    assert(join != null);
 
-    var callbacks = GestureCallbacks(
+    late GestureCallbacks? callbacks;
+    callbacks = GestureCallbacks(
       onTap: onTap,
       onDoubleTap: onDoubleTap,
       onLongPress: onLongPress,
@@ -157,6 +154,8 @@ class GestureCallbacks extends JoinableObject<GestureCallbacks> {
     if (T != dynamic &&
         joinDynamic != null &&
         _global.exists<dynamic>(key: key)) {
+      final dynamicCallbacks = _global.get<dynamic>(key: key);
+      if (callbacks == null) return dynamicCallbacks;
       callbacks = joinDynamic == JoinMethod.combine
           ? callbacks.combine(_global.get<dynamic>(key: key))
           : callbacks.merge(_global.get<dynamic>(key: key));
@@ -167,7 +166,8 @@ class GestureCallbacks extends JoinableObject<GestureCallbacks> {
 
   /// Returns `true` if global callbacks associated with the [key] and
   /// sub-type ([T]) exist.
-  static bool globalCallbacksExist<T>([Key key]) => _global.exists<T>(key: key);
+  static bool globalCallbacksExist<T>([Key? key]) =>
+      _global.exists<T>(key: key);
 
   /// Returns the global callbacks with the corresponding [key]
   /// and sub-type ([T]).
@@ -178,12 +178,14 @@ class GestureCallbacks extends JoinableObject<GestureCallbacks> {
   /// returned callbacks, call [GestureCallbacks.global] to retrieve
   /// callbacks that are joined with the [dynamic] sub-typed callbacks,
   /// if they exist.
-  static GestureCallbacks getGlobalCallbacks<T>([Key key]) =>
+  static GestureCallbacks? getGlobalCallbacks<T>([Key? key]) =>
       _global.get<T>(key: key);
 
   /// Removes the global callbacks with the corresponding [key]
   /// and sub-type ([T]).
-  static GestureCallbacks removeGlobalCallbacks<T>([Key key]) =>
+  ///
+  /// Returns `null` if no corresponding global callbacks exist.
+  static GestureCallbacks? removeGlobalCallbacks<T>([Key? key]) =>
       _global.remove<T>(key: key);
 
   /// Returns a new [GestureCallbacks] by joining [callbacks] with the
@@ -196,9 +198,12 @@ class GestureCallbacks extends JoinableObject<GestureCallbacks> {
   /// If [joinDynamic] isn't `null` and the sub-type ([T]) isn't [dynamic], the
   /// global callbacks associated with the [key] and a [dynamic] sub-type will
   /// be [merge]d or [combine]d into the returned callbacks, if they exist.
-  static GestureCallbacks joinWithGlobal<T>(
-    GestureCallbacks callbacks, {
-    Key key,
+  ///
+  /// Returns `null` if [callbacks] is `null` and no corresponding global
+  /// callbacks exist.
+  static GestureCallbacks? joinWithGlobal<T>(
+    GestureCallbacks? callbacks, {
+    Key? key,
     JoinMethod joinDynamic = JoinMethod.combine,
   }) {
     final globalCallbacks =
